+++ sdk/src/metrics/meter_context_factory.cc
@@ -4,7 +4,6 @@
 #include "opentelemetry/sdk/metrics/meter_context_factory.h"
 #include <memory>
 #include <utility>
-#include <vector>
 #include "opentelemetry/sdk/instrumentationscope/scope_configurator.h"
 #include "opentelemetry/sdk/metrics/meter_config.h"
 #include "opentelemetry/sdk/metrics/meter_context.h"